+++ cio/src/rfds.rs
@@ -457,303 +457,3 @@ impl UpdateAirtableRecord<RFD> for RFD {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::{
-        companies::Company,
-        db::Database,
-        rfds::{clean_rfd_html_links, send_rfd_changelog, update_discussion_link, update_state, NewRFD},
-    };
-
-    #[ignore]
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_monday_cron_rfds_changelog() {
-        crate::utils::setup_logger();
-
-        // Initialize our database.
-        let db = Database::new().await;
-
-        // Get the company id for Oxide.
-        // TODO: split this out per company.
-        let oxide = Company::get_from_db(&db, "Oxide".to_string()).await.unwrap();
-
-        send_rfd_changelog(&db, &oxide).await.unwrap();
-    }
-
-    #[test]
-    fn test_clean_rfd_html_links() {
-        crate::utils::setup_logger();
-
-        let content = r#"https://3.rfd.oxide.computer
-        https://41.rfd.oxide.computer
-        https://543.rfd.oxide.computer#-some-link
-        https://3245.rfd.oxide.computer/things
-        https://3265.rfd.oxide.computer/things
-        <img src="things.png" \>
-        <a href="\#_principles">
-        <object data="thing.svg">
-        <object type="image/svg+xml" data="thing.svg">
-        <a href="\#things" \>
-        link:thing.html[Our thing]
-        link:http://example.com[our example]"#;
-
-        let cleaned = clean_rfd_html_links(content, "0032").unwrap();
-
-        let expected = r#"https://rfd.shared.oxide.computer/rfd/0003
-        https://rfd.shared.oxide.computer/rfd/0041
-        https://rfd.shared.oxide.computer/rfd/0543#-some-link
-        https://rfd.shared.oxide.computer/rfd/3245/things
-        https://rfd.shared.oxide.computer/rfd/3265/things
-        <img src="/static/images/0032/things.png" \>
-        <a href="/rfd/0032#_principles">
-        <object data="/static/images/0032/thing.svg">
-        <object type="image/svg+xml" data="/static/images/0032/thing.svg">
-        <a href="/rfd/0032#things" \>
-        link:https://0032.rfd.oxide.computer/thing.html[Our thing]
-        link:http://example.com[our example]"#;
-
-        assert_eq!(expected, cleaned);
-    }
-
-    #[test]
-    fn test_get_authors() {
-        crate::utils::setup_logger();
-
-        let mut content = r#"sdfsdf
-sdfsdf
-authors: things, joe
-dsfsdf
-sdf
-authors: nope"#;
-        let mut authors = NewRFD::get_authors(content, true).unwrap();
-        let mut expected = "things, joe".to_string();
-        assert_eq!(expected, authors);
-
-        content = r#"sdfsdf
-= sdfgsdfgsdfg
-things, joe
-dsfsdf
-sdf
-:authors: nope"#;
-        authors = NewRFD::get_authors(content, true).unwrap();
-        expected = "".to_string();
-        assert_eq!(expected, authors);
-
-        content = r#"sdfsdf
-= sdfgsdfgsdfg
-things <things@email.com>, joe <joe@email.com>
-dsfsdf
-sdf
-authors: nope"#;
-        authors = NewRFD::get_authors(content, false).unwrap();
-        expected = r#"things <things@email.com>, joe <joe@email.com>"#.to_string();
-        assert_eq!(expected, authors);
-
-        content = r#":authors: Jess <jess@thing.com>
-
-= sdfgsdfgsdfg
-{authors}
-dsfsdf
-sdf"#;
-        authors = NewRFD::get_authors(content, false).unwrap();
-        expected = r#"Jess <jess@thing.com>"#.to_string();
-        assert_eq!(expected, authors);
-    }
-
-    #[test]
-    fn test_get_state() {
-        crate::utils::setup_logger();
-
-        let mut content = r#"sdfsdf
-sdfsdf
-state: discussion
-dsfsdf
-sdf
-authors: nope"#;
-        let mut state = NewRFD::get_state(content).unwrap();
-        let mut expected = "discussion".to_string();
-        assert_eq!(expected, state);
-
-        content = r#"sdfsdf
-= sdfgsdfgsdfg
-:state: prediscussion
-dsfsdf
-sdf
-:state: nope"#;
-        state = NewRFD::get_state(content).unwrap();
-        expected = "prediscussion".to_string();
-        assert_eq!(expected, state);
-    }
-
-    #[test]
-    fn test_get_discussion() {
-        crate::utils::setup_logger();
-
-        let mut content = r#"sdfsdf
-sdfsdf
-discussion: https://github.com/oxidecomputer/rfd/pulls/1
-dsfsdf
-sdf
-authors: nope"#;
-        let mut discussion = NewRFD::get_discussion(content).unwrap();
-        let expected = "https://github.com/oxidecomputer/rfd/pulls/1".to_string();
-        assert_eq!(expected, discussion);
-
-        content = r#"sdfsdf
-= sdfgsdfgsdfg
-:discussion: https://github.com/oxidecomputer/rfd/pulls/1
-dsfsdf
-sdf
-:discussion: nope"#;
-        discussion = NewRFD::get_discussion(content).unwrap();
-        assert_eq!(expected, discussion);
-    }
-
-    #[test]
-    fn test_update_discussion_link() {
-        crate::utils::setup_logger();
-
-        let link = "https://github.com/oxidecomputer/rfd/pulls/2019";
-        let mut content = r#"sdfsdf
-sdfsdf
-discussion:   https://github.com/oxidecomputer/rfd/pulls/1
-dsfsdf
-sdf
-authors: nope"#;
-        let mut result = update_discussion_link(content, link, true);
-        let mut expected = r#"sdfsdf
-sdfsdf
-discussion: https://github.com/oxidecomputer/rfd/pulls/2019
-dsfsdf
-sdf
-authors: nope"#;
-        assert_eq!(expected, result);
-
-        content = r#"sdfsdf
-= sdfgsd
-discussion: fgsdfg
-:discussion: https://github.com/oxidecomputer/rfd/pulls/1
-dsfsdf
-sdf
-:discussion: nope"#;
-        result = update_discussion_link(content, link, false);
-        expected = r#"sdfsdf
-= sdfgsd
-discussion: fgsdfg
-:discussion: https://github.com/oxidecomputer/rfd/pulls/2019
-dsfsdf
-sdf
-:discussion: nope"#;
-        assert_eq!(expected, result);
-
-        content = r#"sdfsdf
-= sdfgsd
-discussion: fgsdfg
-:discussion:
-dsfsdf
-sdf
-:discussion: nope"#;
-        result = update_discussion_link(content, link, false);
-        expected = r#"sdfsdf
-= sdfgsd
-discussion: fgsdfg
-:discussion: https://github.com/oxidecomputer/rfd/pulls/2019
-dsfsdf
-sdf
-:discussion: nope"#;
-        assert_eq!(expected, result);
-    }
-
-    #[test]
-    fn test_update_state() {
-        crate::utils::setup_logger();
-
-        let state = "discussion";
-        let mut content = r#"sdfsdf
-sdfsdf
-state:   sdfsdfsdf
-dsfsdf
-sdf
-authors: nope"#;
-        let mut result = update_state(content, state, true).unwrap();
-        let mut expected = r#"sdfsdf
-sdfsdf
-state: discussion
-dsfsdf
-sdf
-authors: nope"#;
-        assert_eq!(expected, result);
-
-        content = r#"sdfsdf
-= sdfgsd
-state: fgsdfg
-:state: prediscussion
-dsfsdf
-sdf
-:state: nope"#;
-        result = update_state(content, state, false).unwrap();
-        expected = r#"sdfsdf
-= sdfgsd
-state: fgsdfg
-:state: discussion
-dsfsdf
-sdf
-:state: nope"#;
-        assert_eq!(expected, result);
-
-        content = r#"sdfsdf
-= sdfgsd
-state: fgsdfg
-:state:
-dsfsdf
-sdf
-:state: nope"#;
-        result = update_state(content, state, false).unwrap();
-        expected = r#"sdfsdf
-= sdfgsd
-state: fgsdfg
-:state: discussion
-dsfsdf
-sdf
-:state: nope"#;
-        assert_eq!(expected, result);
-    }
-
-    #[test]
-    fn test_get_title() {
-        crate::utils::setup_logger();
-
-        let mut content = r#"things
-# RFD 43 Identity and Access Management (IAM)
-sdfsdf
-title: https://github.com/oxidecomputer/rfd/pulls/1
-dsfsdf
-sdf
-authors: nope"#;
-        let mut title = NewRFD::get_title(content).unwrap();
-        let expected = "Identity and Access Management (IAM)".to_string();
-        assert_eq!(expected, title);
-
-        content = r#"sdfsdf
-= RFD 43 Identity and Access Management (IAM)
-:title: https://github.com/oxidecomputer/rfd/pulls/1
-dsfsdf
-= RFD 53 Bye
-sdf
-:title: nope"#;
-        title = NewRFD::get_title(content).unwrap();
-        assert_eq!(expected, title);
-
-        // Add a test to show what happens for rfd 31 where there is no "RFD" in
-        // the title.
-        content = r#"sdfsdf
-= Identity and Access Management (IAM)
-:title: https://github.com/oxidecomputer/rfd/pulls/1
-dsfsdf
-sdf
-:title: nope"#;
-        title = NewRFD::get_title(content).unwrap();
-        assert_eq!(expected, title);
-    }
-}
